hotfix to logout

+++ src/routes/auth/handlers/log-out.ts
@@ -5,13 +5,14 @@ import prisma from '@/services/prisma'
 export const logOut = async (req: ServerRequest, res: ServerResponse) => {
   const sessionToken = req.signedCookies.session
 
-  console.log('sessionToken', sessionToken)
-
-  await prisma.sessions.deleteMany({
-    where: {
-      session_token: sessionToken,
-    },
-  })
+  // If the sessionToken is undefined, prisma deletes all sessions. Thanks prisma :)
+  if (sessionToken) {
+    await prisma.sessions.deleteMany({
+      where: {
+        session_token: sessionToken,
+      },
+    })
+  }
 
   res.clearCookie('session')
